refactor: move context.ts to context folder
